refactor redux typing
