+++ src/components/pages/profile/form.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useMemo } from 'react';
-import { useGetProfileQuery, usePatchProfileMutation } from '../../../services/apis/auth';
+import { useGetProfileQuery, usePatchProfileMutation } from '../../../services/apis/auth/auth';
 import styles from './inputs.module.css'
 import { useAppDispatch, useAppSelector } from '../../../services/hooks';
-import { setKeyValue, TInputsNames, clearForm } from '../../../services/slices/profileForm';
+import { setKeyValue, TInputsNames, clearForm } from '../../../services/slices/profileForm/profileForm';
 import ProfileInput from './Input';
 import Loader from '../../share/loader/loader';
 import ErrorView from '../../share/error/error';
