+++ src/services/apis/auth/auth.ts
@@ -1,121 +1,10 @@
 import { BaseQueryFn, FetchArgs, FetchBaseQueryError, createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import {lsStorage} from '../../share/browser-storage';
-import { SerializedError } from '@reduxjs/toolkit';
-import { IResponse } from '../../share/typing';
+import {lsStorage} from '../../../share/browser-storage';
+import { IResponseAuthMessage, IResponseMessage, IUserResponse } from './types';
+import { setUserInStorage, getLSUserInfo, refreshTokensInStorage, deleteUserFromStorage, lsUserKeys } from '../../tokens';
+import { apiUrl, jsonHeader, protectedApiRoutes } from '../../constants/varaibles';
+import { baseQuery, check_jwt_expired, getErrorMessage, transformAuthResponse } from '../../utils';
 
-export const apiUrl = 'https://norma.nomoreparties.space/api/'
-
-interface IUserResponse extends IResponse {
-  user: {
-    email: string,
-    name: string
-  }
-}
-
-export interface IResponseMessage extends IResponse {
-  message: string
-}
-
-export interface IResponseAuthMessage extends IUserResponse {
-  accessToken: string,
-  refreshToken: string
-}
-
-export const lsUserKeys = {
-  accessToken: "accessToken",
-  refreshToken: "refreshToken",
-  username: "username",
-  email: "email"
-}
-
-export const getLSUserInfo: () => {
-  name: string;
-  email: string;
-  } | undefined = () => {
-  const name = lsStorage.get(lsUserKeys.username)
-  const email = lsStorage.get(lsUserKeys.email)
-  if(name && email){
-    return {name, email}
-  }
-}
-
-export const refreshTokensInStorage = (accessToken: string, refreshToken: string):void => {
-  lsStorage.set(lsUserKeys.accessToken, accessToken.split("Bearer ")[1])
-  lsStorage.set(lsUserKeys.refreshToken, refreshToken)
-}
-
-const setUserInStorage = ({user, accessToken, refreshToken}:IResponseAuthMessage):void => {
-  lsStorage.set(lsUserKeys.email, user.email)
-  lsStorage.set(lsUserKeys.username, user.name)
-  refreshTokensInStorage(accessToken, refreshToken)
-}
-
-export const deleteUserFromStorage = ():void => Object.keys(lsUserKeys).forEach( key => lsStorage.delete(key))
-
-export const getErrorMessage = (error: FetchBaseQueryError | SerializedError | undefined): string | null => {
-  if(!error) return null
-  if('data' in error && typeof error.data === 'object' && 
-    error.data && 'message' in error.data){
-    return error.data?.message as string
-  }
-  return null
-}
-
-const jwt_expired_403:string = "jwt expired"
-const jwt_malformed_403:string = "jwt malformed"
-
-export const check_jwt_expired = (error: FetchBaseQueryError):boolean => {
-  if(!error) return false
-  if(error.status === 401) return true
-  const error_message = getErrorMessage(error)
-  if(error.status === 403 && error_message && [
-    jwt_expired_403, jwt_malformed_403
-  ].includes(error_message)) {
-    return true
-  }
-  return false
-}
-
-export const transformAuthResponse = (
-  response:IResponseAuthMessage, 
-  messages:{
-    success_message: string,
-    error_message: string
-  }):IResponseMessage => {
-  if(response.success){
-    setUserInStorage(response)
-    return {
-      success: response.success,
-      message: messages.success_message
-    }
-  } else {
-    return {
-      success: response.success,
-      message: messages.error_message
-    }
-  }
-}
-
-const jsonHeader:{[key:string]: string} = {
-  'Content-Type': 'application/json',
-}
-
-const baseQuery = fetchBaseQuery({ 
-  baseUrl: apiUrl,
-  prepareHeaders: (headers) => {
-    const token = lsStorage.get(lsUserKeys.accessToken)
-    if (token) {
-      headers.set('Authorization', `Bearer ${token}`);
-    }
-    return headers;
-  },
-})
-
-const protectedApiRoutes:string[] = [
-  "auth/user",
-  "auth/logout",
-  "orders"
-]
 
 export const getReauthBaseQuery = (_baseQuery:BaseQueryFn<
   string | FetchArgs,
