+++ src/services/slices/burger/types.ts
@@ -1,0 +1,7 @@
+import { IIngredient, ConstructorIngredient } from "../../../share/typing"
+
+export interface IBurgerState {
+    bun: IIngredient|null,
+    core: ConstructorIngredient[]
+}
+
