+++ src/services/slices/tabs/types.ts
@@ -1,0 +1,6 @@
+export interface ITabState {
+    elementsTop: {
+      [key:string]: number
+    }
+    containerScrolTop: number
+}
