+++ src/services/slices/user/types.ts
@@ -1,0 +1,7 @@
+export interface IUserState {
+    authorized: boolean,
+    loading: boolean,
+    error: boolean,
+    name?: string|undefined,
+    email?: string|undefined
+}
