+++ src/share/typing.ts
@@ -20,6 +20,10 @@ export interface IIngredient {
     __v: number
 }
 
+export type ISODateString = string
+export const isISODateString = (dateString: string): dateString is ISODateString => /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z$/.test(dateString)
+
+
 export interface ConstructorIngredient extends IIngredient {
     uid: string
 }
